+++ src/API/Common/IFactory.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace \API\Common;
+namespace API\Common;
 
 interface IFactory
 {
