+++ src/API/Mail/Classes/Mail.php
@@ -1,5 +1,5 @@
 <?php  
-namespace \API\Mail;
+namespace API\Mail;
 
 class Mail implements IMail
 {
