+++ src/API/Mail/Classes/MailAddress.php
@@ -1,5 +1,5 @@
 <?php  
-namespace \API\Mail;
+namespace API\Mail;
 
 class MailAddress implements IMailAddress
 {
