+++ src/API/Mail/Classes/Message.php
@@ -1,5 +1,5 @@
 <?php  
-namespace \API\Mail;
+namespace API\Mail;
 
 class Message implements IMessage
 {
