+++ src/API/Mail/Interfaces/IAddressee.php
@@ -1,5 +1,5 @@
 <?php  
-namespace \API\Mail;
+namespace API\Mail;
 
 interface IAddressee
 {
