+++ src/API/Mail/Interfaces/IHeader.php
@@ -1,5 +1,5 @@
 <?php  
-namespace \API\Mail;
+namespace API\Mail;
 
 interface IHeader
 {
