+++ src/API/Send/AMailSender.php
@@ -1,9 +1,10 @@
 <?php
 
-use \API\Common;
-use \API\Tools;
+namespace API\Send;
 
-namespace \API\Send;
+use API\Common;
+use API\Tools;
+
 /**
  * supported interfaces:
  * - API.SendIMailSender
