+++ src/API/Send/IMailSender.php
@@ -1,9 +1,9 @@
 <?php
 
-use \API\Common;
-use \API\Tools;
+namespace API\Send;
 
-namespace \API\Send;
+use API\Common;
+use API\Tools;
 
 interface IMailSender extends ICreateInstance
 {
