+++ src/API/Send/MailSenderNot4Factory.php
@@ -1,9 +1,10 @@
 <?php
 
-use \API\Common;
-use \API\Tools;
+namespace API\Send;
 
-namespace \API\Send;
+use API\Common;
+use API\Tools;
+
 /**
  * All classes are inherits from this not will register on factory
  */
