+++ src/API/Tools/Classes/EventsHub.php
@@ -1,5 +1,5 @@
 <?php  
-namespace \API\Tools;
+namespace API\Tools;
 
 
 /**
